Use either "unsigned char *" (for pointers in generated code) or "int" (for linear DOS space addresses) but no more long casts for consistency between i386 and x86-64.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1976 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/emu-i386/simx86/sigsegv.c
@@ -530,7 +530,7 @@ int e_emu_fault(struct sigcontext_struct *scp)
 		    return 1;
 		/* We HAVE to invalidate all the code in the page
 		 * if the page is going to be unprotected */
-		InvalidateNodePage(_cr2, 0, (long)p, &codehit);
+		InvalidateNodePage(_cr2, 0, p, &codehit);
 		e_resetpagemarks((void *)_cr2, 1);
 		e_munprotect((void *)_cr2, 0);
 		/* now go back and perform the faulting op */
